+++ test/test_integration.py
@@ -32,7 +32,7 @@ import unittest
 import grpc
 import numpy as np
 import philote_mdo.general as pmdo
-from philote_mdo.examples import Paraboloid
+from philote_mdo.examples import Paraboloid, QuadradicImplicit
 
 
 class IntegrationTests(unittest.TestCase):
@@ -42,7 +42,7 @@ class IntegrationTests(unittest.TestCase):
 
     def test_paraboloid_compute(self):
         """
-        Tests the compute function of the Paraboloid server.
+        Integration test for the Paraboloid compute function.
         """
         # server code
         server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
@@ -72,6 +72,43 @@ class IntegrationTests(unittest.TestCase):
         outputs = client.run_compute(inputs)
 
         self.assertEqual(outputs["f_xy"][0], 39.0)
+
+        # end the server
+        server.stop(0)
+
+    def test_quadratic_compute_residuals(self):
+        """
+        Integration test for the QuadraticImplicit compute function.
+        """
+        # server code
+        server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
+
+        discipline = pmdo.ImplicitServer(discipline=QuadradicImplicit())
+        discipline.attach_to_server(server)
+
+        server.add_insecure_port("[::]:50051")
+        server.start()
+
+        # client code
+        client = pmdo.ImplicitClient(channel=grpc.insecure_channel("localhost:50051"))
+
+        # transfer the stream options to the server
+        client.send_stream_options()
+
+        # run setup
+        client.run_setup()
+        client.get_variable_definitions()
+        client.get_partials_definitions()
+
+        # define some inputs
+        inputs = {"a": np.array([1.0]), "b": np.array([2.0]), "c": np.array([2.0])}
+        outputs = {"x": np.array([1.0])}
+        residuals = {}
+
+        # run a function evaluation
+        residuals = client.run_compute_residuals(inputs, outputs)
+
+        self.assertEqual(residuals["x"][0], 5.0)
 
         # end the server
         server.stop(0)